requirements.txt
asgiref==3.6.0
cffi==1.15.1
contextlib2==21.6.0
cryptography==39.0.1
Django==4.1.7
django-ninja==0.21.0
django-ninja-extra==0.18.6
django-ninja-jwt==5.2.2
dnspython==2.3.0
email-validator==1.3.1
idna==3.4
injector==0.20.1
ninja-schema==0.13.0
orjson==3.8.7
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
sqlparse==0.4.3
typing_extensions==4.5.0
